+++ lang/da/fields.php
@@ -1,11 +1,14 @@
 <?php
 
 return [
-    'name' => 'Navn',
-    'email' => 'E-mail',
-    'email_contact' => 'Efterlad din e-mail til kontakt',
-    'message' => 'Beskriv dit problem',
-    'phone' => 'Telefon',
+
+    'email' => 'Din e-mailadresse',
+    'email_optional' => 'Efterlad din e-mail for kontakt',
+    'fullname' => 'Dit fulde navn',
+    'message' => 'Skriv din besked her',
+    'name' => 'Dit fulde navn',
     'optional' => 'valgfrit',
+    'phone' => 'Dit telefonnummer',
     'submit' => 'Send',
+
 ];