+++ lang/da/layouts.php
@@ -1,22 +1,16 @@
 <?php
 
 return [
-    'modal.title' => 'Send feedback',
+
+    'email.anonymous' => '-',
+    'email.from' => 'Fra',
+    'email.no_message' => 'Ingen besked',
+    'email.no_subject' => 'Intet emne',
+    'email.sent_from' => 'Sendt fra',
+    'email.subject' => 'Feedback :app_www',
+    'modal.description' => 'Din feedback hjælper os med at forbedre. Del venligst dine tanker.',
     'modal.link_text' => 'Feedback',
     'modal.success_message' => 'Tak for din feedback!',
+    'modal.title' => 'Send feedback',
 
-    'form.name' => 'Navn',
-    'form.email' => 'E-mail',
-    'form.phone' => 'Telefon',
-    'form.note' => 'Besked',
-    'form.message' => 'Beskriv dit problem',
-    'form.optional' => 'valgfrit',
-    'form.submit' => 'Send feedback',
-
-    'email.subject' => 'Feedback :app_www',
-    'email.no_subject' => 'Intet emne',
-    'email.no_message' => 'Ingen besked',
-    'email.from' => 'Fra',
-    'email.anonymous' => '-',
-    'email.send_from' => 'Sendt fra',
 ];