+++ lang/da/messages.php
@@ -2,4 +2,5 @@
 
 return [
     'success' => 'Tak for din feedback!',
+    'phone_validation' => 'Indtast venligst et gyldigt telefonnummer med 7-15 cifre.',
 ];