+++ resources/views/emails/feedback-plain.blade.php
@@ -3,10 +3,10 @@
 @section('content')
 @foreach($feedback as $field)
 {{ $field['label'] }}:
-{{ $field['value'] }}
+{!! html_entity_decode($field['value']) !!}
 
 @endforeach
-@lang('ig-feedback::layouts.email.send_from')
+@lang('feedback::layouts.email.sent_from'):
 {{ session('currentPage') ?? '-' }}
 {{ InternetGuru\LaravelCommon\Support\Helpers::getAppInfo() }}
 @endsection