+++ resources/views/livewire/feedback.blade.php
@@ -14,9 +14,13 @@
                             <button type="button" class="btn-close" wire:click="closeModal" aria-label="Close"></button>
                         </div>
 
+                        @if($description)
+                            <p class="me-5">{{ $description }}</p>
+                        @endif
+
                         @if($showSuccess)
                             <div class="alert alert-success" role="alert">
-                                {{ __('ig-feedback::messages.success') }}
+                                {{ __('feedback::messages.success') }}
                             </div>
                         @else
                             <form wire:submit.prevent="send" class="editable-skip">
@@ -44,7 +48,7 @@
                                 @endforeach
 
                                 <x-ig::submit>
-                                    {{ $submit ?? __('ig-feedback::fields.submit') }}
+                                    {{ $submit ?? __('feedback::fields.submit') }}
                                 </x-ig::submit>
                             </form>
                         @endif