+++ CHANGELOG.md
@@ -4,6 +4,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [4.1.0] - 2025-11-09
+
+_Stable release based on [4.1.0-rc.1]._
+
+## [4.1.0-rc.1] - 2025-11-07
+
+### Added
+
+- Add button and link trigger components.
+- Add an optional description attribute.
+- Add phone number validation.
+
+### Changed
+
+- Set a specific default label for optional email.
+- Update translations and remove unused.
+
+### Fixed
+
+- Fix value contents in email body.
+
 ## [4.0.2] - 2025-11-07
 
 ### Fixed
@@ -239,6 +260,8 @@ _Stable release based on [0.1.0-rc.1]._
 
 - New changelog file.
 
+[4.1.0]: https://https://github.com/internetguru/laravel-feedback/compare/v4.0.2...v4.1.0
+[4.1.0-rc.1]: https://github.com/internetguru/laravel-feedback/releases/tag/v4.0.2
 [4.0.2]: https://https://github.com/internetguru/laravel-feedback/compare/v4.0.1...v4.0.2
 [4.0.1]: https://https://github.com/internetguru/laravel-feedback/compare/v4.0.0...v4.0.1
 [4.0.0]: https://https://github.com/internetguru/laravel-feedback/compare/v3.1.3...v4.0.0